Create README.md
Create the external style sheet for the home page from the main.css file that's in the style sheet folder. Set the default  font for the document to san-serif font. Adjust the font sizes for all of the headings. In addition, format the first letter of the first heading in the main content so it’s larger than the other letters in the heading. In the header, float the image to the left and indent both headings. In addition, change the color of the first heading to orange, italicize it, and add a black shadow. Format the links so they’re displayed in boldfaced orange whether or not they’ve been visited. If a link has the focus or the mouse is hovering over it, though, it should be 
displayed in green. Format the list to increase the space between the list items.  Reduce the font size for the footer and center it.  Link the main style sheet that you just created to the home page.